# docs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: DPO оператора
  description: "REST API центра управления согласиями субъектов персональных данны\
    х. Идентификаторы — на английском, пользовательские тексты — на русском (NFR-8)."
  license:
    name: Proprietary
  title: In consensu — центр управления согласиями
  version: v1
servers:
- description: Текущий хост
  url: /
paths:
...
  /api/v1/consents:
    get:
      operationId: list_6
      parameters:
      - in: query
        name: subjectId
        required: false
        schema:
          type: string
          format: uuid
      - in: query
        name: typeCode
        required: false
        schema:
          type: string
      - in: query
        name: status
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - EXPIRING
          - EXPIRED
          - REVOKED
          - SUPERSEDED
      - in: query
        name: thirdPartyId
        required: false
        schema:
          type: string
          format: uuid
      - in: query
        name: source
        required: false
        schema:
          type: string
          enum:
          - CLIENT_BASE_IMPORT
          - CONTRACT
          - SUPPLEMENTARY_AGREEMENT
          - PERSONAL_ACCOUNT_REGISTRATION
          - WEBSITE_APPLICATION
          - LOYALTY_PROGRAM
          - MOBILE_APP
          - CALL_CENTER
          - OFFICE
          - OTHER
      - in: query
        name: validUntilFrom
        required: false
        schema:
          type: string
          format: date-time
      - in: query
        name: validUntilTo
        required: false
        schema:
          type: string
          format: date-time
      - in: query
        name: pageable
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseConsentResponse"
          description: OK
      tags:
      - consent-controller
    post:
      operationId: register
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterConsentRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RegisterConsentResponse"
          description: OK
      tags:
      - consent-controller
  /api/v1/consents/{id}:
    get:
      operationId: get_8
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConsentResponse"
          description: OK
      tags:
      - consent-controller
  /api/v1/consents/{id}/evidence:
    get:
      operationId: evidence
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConsentEvidenceResponse"
          description: OK
      tags:
      - consent-controller
  /api/v1/consents/{id}/revocation-preview:
    get:
      operationId: revocationPreview
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsentResponse"
          description: OK
      tags:
      - consent-controller
  /api/v1/consents/{id}/revoke:
    post:
      operationId: revoke_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevocationResponse"
          description: OK
      tags:
      - consent-controller
...
  /api/v1/forms/{id}/preview:
    get:
      operationId: preview
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
          description: OK
      tags:
      - consent-form-controller
...
  /api/v1/self/consents/{id}/revoke:
    post:
      operationId: revoke
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: externalId
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevocationReceipt"
          description: OK
      tags:
      - self-service-controller
...
components:
  schemas:
...
    ConsentEvidenceResponse:
      type: object
      properties:
        consentId:
          type: string
          format: uuid
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventEntry"
        form:
          $ref: "#/components/schemas/FormText"
        integrity:
          type: string
        integrityProblems:
          type: array
          items:
            type: string
        signature:
          $ref: "#/components/schemas/Signature"
...
    ConsentResponse:
      type: object
      properties:
        daysLeft:
          type: integer
          format: int64
        form:
          $ref: "#/components/schemas/FormRef"
        grantedAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        pdnCategories:
          type: array
          items:
            type: string
        purposes:
          type: array
          items:
            type: string
        revocationSource:
          type: string
        revokedAt:
          type: string
          format: date-time
        signatureType:
          type: string
        source:
          type: string
        sourceRef:
          type: string
        status:
          type: string
        statusText:
          type: string
        subjectId:
          type: string
          format: uuid
        thirdPartyId:
          type: string
          format: uuid
        thirdPartyName:
          type: string
        typeCode:
          type: string
        typeId:
          type: string
          format: uuid
        typeNameRu:
          type: string
        validUntil:
          type: string
          format: date-time
...
    ContactPayload:
      type: object
      properties:
        primary:
          type: boolean
        type:
          type: string
          enum:
          - PHONE
          - EMAIL
          - POSTAL_ADDRESS
        value:
          type: string
      required:
      - type
...
    EventEntry:
      type: object
      properties:
        actor:
          type: string
        eventType:
          type: string
        eventTypeRu:
          type: string
        hash:
          type: string
        occurredAt:
          type: string
          format: date-time
...
    FormRef:
      type: object
      properties:
        checksum:
          type: string
        code:
          type: string
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
...
    FormText:
      type: object
      properties:
        checksumMatches:
          type: boolean
        code:
          type: string
        publishedAt:
          type: string
          format: date-time
        recalculatedChecksum:
          type: string
        storedChecksum:
          type: string
        text:
          type: string
        version:
          type: integer
          format: int32
...
    ItemPayload:
      type: object
      properties:
        accepted:
          type: boolean
        formItemId:
          type: string
          format: uuid
      required:
      - formItemId
...
    PageResponseConsentResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ConsentResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    RegisterConsentRequest:
      type: object
      properties:
        evidence:
          type: object
          additionalProperties:
            type: object
        formId:
          type: string
          format: uuid
        grantedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemPayload"
        signatureType:
          type: string
          enum:
          - SIMPLE_ES_SMS
          - SIMPLE_ES_LK
          - HANDWRITTEN
          - UKEP
          - IMPORTED_LEGACY
        source:
          type: string
          enum:
          - CLIENT_BASE_IMPORT
          - CONTRACT
          - SUPPLEMENTARY_AGREEMENT
          - PERSONAL_ACCOUNT_REGISTRATION
          - WEBSITE_APPLICATION
          - LOYALTY_PROGRAM
          - MOBILE_APP
          - CALL_CENTER
          - OFFICE
          - OTHER
        sourceRef:
          type: string
        subject:
          $ref: "#/components/schemas/SubjectPayload"
        subjectExternalId:
          type: string
      required:
      - formId
      - items
      - signatureType
      - source
    RegisterConsentResponse:
      type: object
      properties:
        consents:
          type: array
          items:
            $ref: "#/components/schemas/ConsentResponse"
        declinedItems:
          type: array
          items:
            type: string
            format: uuid
        idempotentReplay:
          type: boolean
...
    RevocationReceipt:
      type: object
      properties:
        caseNumber:
          type: string
        message:
          type: string
        revokedAt:
          type: string
          format: date-time
        revokedConsentIds:
          type: array
          items:
            type: string
            format: uuid
    RevocationResponse:
      type: object
      properties:
        cascadedConsentIds:
          type: array
          items:
            type: string
            format: uuid
        caseNumber:
          type: string
        consentId:
          type: string
          format: uuid
        processingStopDeadline:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
    RevokeRequest:
      type: object
      properties:
        caseNumber:
          type: string
        evidence:
          type: object
          additionalProperties:
            type: object
        reason:
          type: string
        revocationSource:
          type: string
          enum:
          - PERSONAL_ACCOUNT
          - MOBILE_APP
          - WRITTEN_REQUEST
          - CALL_CENTER
          - EMAIL_REQUEST
          - CASCADE
      required:
      - caseNumber
      - reason
      - revocationSource
...
    Signature:
      type: object
      properties:
        evidence:
          type: object
          additionalProperties:
            type: object
        signatureType:
          type: string
        signatureTypeRu:
          type: string
...
    SubjectPayload:
      type: object
      properties:
        birthDate:
          type: string
          format: date
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/ContactPayload"
        externalId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
...